extend to accept multiple files to check
This is useful for integration into tools like
treefmt (https://github.com/numtide/treefmt).

Currently we only consider the first arguments for gitignore, which
should work if people run on this tool on their repository. If multiple
arguments are provided they probably want to be any of these targets to
be checked.